+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 * Added support for querying inverse relationships (#2904).
 * Moved inverse relationships out of beta stage.
 * Added `Realm.getDefaultConfiguration()` (#4725).
+* Added `RealmObject.getConfiguration(RealmModel)` (#4720).
 
 ### Bug Fixes